+++ lib/ui/pages/home_page/components/loading_screen.dart
