+++ lib/ui/pages/home_page/components/search_item_card.dart
