+++ lib/ui/pages/pages.dart
@@ -37,27 +37,27 @@ part 'detail_menu_page/components/quantity_button.dart';
 
 part 'detail_menu_page/components/add_to_basket_button.dart';
 
-part 'restaurant_home_page/restaurant_home_page.dart';
+part 'home_page/home_page.dart';
 
-part 'restaurant_home_page/components/appbar.dart';
+part 'home_page/components/appbar.dart';
 
-part 'restaurant_home_page/components/menu_item_list.dart';
+part 'home_page/components/menu_item_list.dart';
 
-part 'restaurant_home_page/components/fab_checkout.dart';
+part 'home_page/components/fab_checkout.dart';
 
-part 'restaurant_home_page/components/loading_screen.dart';
+part 'home_page/components/loading_screen.dart';
 
-part 'restaurant_home_page/components/RecommendationMenu.dart';
+part 'home_page/components/RecommendationMenu.dart';
 
-part 'restaurant_home_page/components/restaurant_info.dart';
+part 'home_page/components/restaurant_info.dart';
 
-part 'restaurant_home_page/components/category_header_list.dart';
+part 'home_page/components/category_header_list.dart';
 
 part 'payment_method_page/payment_method_page.dart';
 
 part 'order_status_page/order_status_page.dart';
 
-part 'restaurant_home_page/components/search_item_card.dart';
+part 'home_page/components/search_item_card.dart';
 
 part 'scan_qr_page/scan_qr_code.dart';
