+++ lib/utils/number_format.dart
@@ -1,7 +1,9 @@
 part of 'utils.dart';
 
-String currency(double number) => NumberFormat.currency(
-      locale: 'id',
-      decimalDigits: 0,
-      name: '',
-    ).format(number);
+String currency(double number) {
+  return NumberFormat.currency(
+    locale: 'id',
+    decimalDigits: 0,
+    name: '',
+  ).format(number);
+}
