--- lib/utils/urlauncher.dart
@@ -1,9 +1,0 @@
-part of 'utils.dart';
-
-launchURL(String url) async {
-  // if (await canLaunch(url)) {
-    await launch(url);
-  // } else {
-  //   throw 'Could not launch $url';
-  // }
-}
