+++ lib/utils/utils.dart
@@ -4,12 +4,7 @@ import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 import 'package:bangorder_mobile/providers/providers.dart';
 import 'package:bangorder_mobile/models/models.dart';
-import 'package:url_launcher/url_launcher.dart';
-
-part 'call_services.dart';
 
 part 'number_format.dart';
 
 part 'data_search.dart';
-
-part 'urlauncher.dart';
